react makes fetch request to express.js server
