+++ client/src/solver/BlockRowCol.js
