+++ client/src/solver/CheckPuzzle.js
