+++ client/src/solver/SoleCandidate.js
