+++ client/src/solver/Solver.js
