+++ client/src/solver/UniqueCandidate.js
