+++ client/src/solver/Utility.js
